Moved ik to shared folder
