+++ src/js/shared/IK/ResourceManager.js
