+++ src/js/shared/IK/objects/BoneRotationControl.js
