+++ src/js/shared/IK/objects/IkObjects/ChainObject.js
